+++ src/components/ProductImage/ProductImage.tsx
@@ -4,16 +4,17 @@ import styles from './ProductImage.module.css'
 interface Props {
   name: string
   image: string
+  isConnected: boolean
 }
 
 // TODO: Use typography here! and change the CSS names
 
-export function ProductImage({ name, image }: Props) {
+export function ProductImage({ name, image, isConnected }: Props) {
   return (
     <div className={styles.container}>
       <div className={styles.header}>
         <div className={styles.title}>{name}</div>
-        <ConnectionIndicator isConnected />
+        <ConnectionIndicator isConnected={isConnected} />
       </div>
       <div className={styles.imageContainer}>
         <img src={image} alt={name} className={styles.image} />